Consume SmallRye Config 2.1.0 and tidy FATs
