+++ dev/cnf/dependabot/check_this_in_if_it_changes/pom.xml
@@ -524,17 +524,17 @@
     <dependency>
       <groupId>io.smallrye.config</groupId>
       <artifactId>smallrye-config-common</artifactId>
-      <version>2.0.2</version>
+      <version>2.1.0</version>
     </dependency>
     <dependency>
       <groupId>io.smallrye.config</groupId>
       <artifactId>smallrye-config-core</artifactId>
-      <version>2.0.2</version>
+      <version>2.1.0</version>
     </dependency>
     <dependency>
       <groupId>io.smallrye.config</groupId>
       <artifactId>smallrye-config</artifactId>
-      <version>2.0.2</version>
+      <version>2.1.0</version>
     </dependency>
     <dependency>
       <groupId>io.smallrye</groupId>
