+++ dev/io.openliberty.microprofile.config.2.0.internal_fat/fat/src/io/openliberty/microprofile/config/internal_fat/apps/brokenInjection/BadConfigPropertyInMethodBean.java
@@ -16,7 +16,7 @@ import javax.inject.Inject;
 import org.eclipse.microprofile.config.inject.ConfigProperty;
 
 @RequestScoped
-public class ConfigUnnamedMethodInjectionBean {
+public class BadConfigPropertyInMethodBean {
 
     @Inject
     public void aMethod(@ConfigProperty String aMethodParameter) {
