+++ dev/io.openliberty.microprofile.config.2.0.internal_fat/publish/servers/Config20ExceptionServer/bootstrap.properties
@@ -1,3 +1,1 @@
 bootstrap.include=../testports.properties
-osgi.console=7773
-org.jboss.logging.provider=slf4j
