+++ dev/io.openliberty.microprofile.config.2.0.internal_fat/publish/servers/Config20Server/bootstrap.properties
@@ -1,4 +1,2 @@
 bootstrap.include=../testports.properties
-osgi.console=7773
-org.jboss.logging.provider=slf4j
 KeyDefinedIn2Places=valueFromBootstrapProperties
